refactor: Extract context-free tools into AISystemTools class
Context-free tools (getCurrentDateTime, getCurrentUser,
getCurrentUserProfiles) were in AIObjectTools which implements
iAIContextAwareToolProvider. This caused them to be unavailable
in conversations without an object context.

- New AISystemTools class (iAIToolProvider only) for the 3 tools
- AIObjectTools now contains only 6 context-dependent tools
- AIService categorizes tools into always-available vs context-dependent
- ContinueConversation $aTools default changed from [] to null
  (null=auto, []=disable, array=explicit)
- System tools are now available in every conversation
