+++ src/Helper/AIObjectTools.php
@@ -33,13 +33,15 @@ use Itomig\iTop\Extension\AIBase\Contracts\iAIToolProvider;
 use LLPhant\Chat\FunctionInfo\FunctionInfo;
 use LLPhant\Chat\FunctionInfo\Parameter;
 use MetaModel;
-use UserRights;
 
 /**
  * Default AI tools for interacting with iTop DBObjects.
  *
  * These tools are read-only and safe for AI use. They provide access to
- * object properties and attribute information.
+ * object properties and attribute information. All tools require an object context.
+ *
+ * Context-free tools (getCurrentDateTime, getCurrentUser, getCurrentUserProfiles)
+ * have been moved to AISystemTools.
  *
  * Lifecycle-specific tools (getState, getStateLabel, getAvailableTransitions) are
  * available as methods but not registered as default AI tools, since not all
@@ -220,106 +222,6 @@ class AIObjectTools implements iAIToolProvider, iAIContextAwareToolProvider
 	}
 
 	/**
-	 * Get the current date and time.
-	 *
-	 * @return string Current date and time in ISO 8601 format.
-	 */
-	public function getCurrentDateTime(): string
-	{
-		IssueLog::Debug(__METHOD__ . ": Called (no context needed)", AIBaseHelper::MODULE_CODE);
-		$result = date('Y-m-d H:i:s');
-		IssueLog::Debug(__METHOD__ . ": Returning: " . $result, AIBaseHelper::MODULE_CODE);
-		return $result;
-	}
-
-	/**
-	 * Get information about the currently logged-in user.
-	 *
-	 * @return string JSON with user info (id, login, language, contact details).
-	 */
-	public function getCurrentUser(): string
-	{
-		IssueLog::Debug(__METHOD__ . ": Called (no context needed)", AIBaseHelper::MODULE_CODE);
-		$oUser = UserRights::GetUserObject();
-		if ($oUser === null) {
-			return json_encode(['error' => 'No user logged in']);
-		}
-
-		$aResult = [
-			'user' => [
-				'id' => (string) $oUser->GetKey(),
-				'login' => $oUser->Get('login'),
-				'language' => $oUser->Get('language'),
-				'contact' => null,
-			],
-		];
-
-		$iContactId = (int) $oUser->Get('contactid');
-		if ($iContactId > 0) {
-			$oContact = MetaModel::GetObject('Contact', $iContactId, false);
-			if ($oContact !== null) {
-				$aContact = [
-					'id' => (string) $oContact->GetKey(),
-					'class' => get_class($oContact),
-					'friendlyname' => $oContact->GetName(),
-					'name' => $oContact->Get('name'),
-					'email' => $oContact->Get('email'),
-					'org_id' => (string) $oContact->Get('org_id'),
-					'org_name' => $oContact->Get('org_name'),
-				];
-
-				// Person hat zusätzlich first_name
-				if ($oContact instanceof \Person) {
-					$aContact['first_name'] = $oContact->Get('first_name');
-				}
-
-				$aResult['user']['contact'] = $aContact;
-			}
-		}
-
-		IssueLog::Debug(__METHOD__ . ": Returning user info for " . $oUser->Get('login'), AIBaseHelper::MODULE_CODE);
-		return json_encode($aResult, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE);
-	}
-
-	/**
-	 * Get the permission profiles (roles) assigned to the currently logged-in user.
-	 *
-	 * @return string JSON with user_id, login, and array of profiles with name and description.
-	 */
-	public function getCurrentUserProfiles(): string
-	{
-		IssueLog::Debug(__METHOD__ . ": Called (no context needed)", AIBaseHelper::MODULE_CODE);
-		$oUser = UserRights::GetUserObject();
-		if ($oUser === null) {
-			return json_encode(['error' => 'No user logged in']);
-		}
-
-		$aProfiles = UserRights::ListProfiles();
-		$aProfileDetails = [];
-
-		foreach ($aProfiles as $iProfileId => $sProfileName) {
-			$aProfileInfo = [
-				'name' => $sProfileName,
-				'description' => '',
-			];
-			$oProfile = MetaModel::GetObject('URP_Profiles', $iProfileId, false);
-			if ($oProfile !== null) {
-				$aProfileInfo['description'] = $oProfile->Get('description');
-			}
-			$aProfileDetails[] = $aProfileInfo;
-		}
-
-		$aResult = [
-			'user_id' => (string) $oUser->GetKey(),
-			'login' => $oUser->Get('login'),
-			'profiles' => $aProfileDetails,
-		];
-
-		IssueLog::Debug(__METHOD__ . ": Returning " . count($aProfileDetails) . " profiles for " . $oUser->Get('login'), AIBaseHelper::MODULE_CODE);
-		return json_encode($aResult, JSON_PRETTY_PRINT | JSON_UNESCAPED_UNICODE);
-	}
-
-	/**
 	 * Get a JSON schema describing the current object's class and all its attributes.
 	 *
 	 * Returns attribute codes, labels, types, and descriptions so the LLM can
@@ -454,30 +356,9 @@ class AIObjectTools implements iAIToolProvider, iAIContextAwareToolProvider
 				[new Parameter('attributeCode', 'string', 'The attribute code')]
 			),
 			new FunctionInfo(
-				'getCurrentDateTime',
-				$this,
-				'Get current server date and time. No parameters required.',
-				[],
-				[]
-			),
-			new FunctionInfo(
 				'describeObject',
 				$this,
 				'Get a JSON schema describing the current object\'s class, all attributes with their codes, labels, types, and descriptions. Call this to discover which attribute codes are available for getAttribute(). No parameters required.',
-				[],
-				[]
-			),
-			new FunctionInfo(
-				'getCurrentUser',
-				$this,
-				'Get information about the currently logged-in user: user ID, login, language, and linked contact/person details (name, email, organization). No parameters required.',
-				[],
-				[]
-			),
-			new FunctionInfo(
-				'getCurrentUserProfiles',
-				$this,
-				'Get the permission profiles (roles) assigned to the currently logged-in user. Each profile has a name and description explaining what permissions it grants. No parameters required.',
 				[],
 				[]
 			),
